Commit classwork 2019-11-28

+++ kcpawan/PlayWithDB/Model.cs
@@ -4,8 +4,10 @@ using Microsoft.EntityFrameworkCore;
 
 public class CompanyContext : DbContext 
 {
+    //internal object Department;
+
     public DbSet<Employee> Employees{get; set;}
-    public DbSet<Department> Departments {get; set;}
+    public DbSet<Department> Departments{get; set;}
     protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
      => optionsBuilder.UseSqlite("Data Source = Company.db");   
 }
@@ -15,7 +17,7 @@ public class Employee
     public int ID {get; set;}
     public string Designation{get; set;}
     public string Name{get; set;}
-    public decimal Salary{get; set;}
+    public double Salary{get; set;}
     public DateTime DOB{get; set;}
     public int DepartmentId{get; set;}
